fix: dimentions race condition
